requirements.txt
Flask
Pillow
pillow-heif
pyinstaller